Makefile, project.scala, lexer.scala, parser.scala: Updated Makefile to allow full use of Scala Metals, fixed typo in Lexer and cleared Parser

+++ src/main/scala/lexer.scala
@@ -107,7 +107,7 @@ object lexer {
     val stringType: Parsley[Unit] = string("string").as()
 
     val identifier: Parsley[String] = lexer.lexeme.names.identifier
-    val integers: Parsley[BigInt] = lexer.lexeme.unsigned.decimal32
+    val integers: Parsley[Int] = lexer.lexeme.unsigned.decimal32
     val charLiterals: Parsley[Char] = lexer.lexeme.character.ascii
     val stringLiterals: Parsley[String] = lexer.lexeme.string.ascii
     val boolLiterals: Parsley[Boolean] = string("true").as(true) <|> string("false").as(false)
